Processing changed to float black level

+++ src/processing/raw_processing.h
@@ -229,10 +229,10 @@ void processingFindWhiteBalance(processingObject_t * processing,
 
 
 /* Black/white level set */
-void processingSetBlackLevel(processingObject_t * processing, int mlvBlackLevel, int mlvBitDepth);
+void processingSetBlackLevel(processingObject_t * processing, float mlvBlackLevel, int mlvBitDepth);
 void processingSetWhiteLevel(processingObject_t * processing, int mlvWhiteLevel, int mlvBitDepth);
-void processingSetBlackAndWhiteLevel( processingObject_t * processing,
-                                      int mlvBlackLevel, int mlvWhiteLevel, int mlvBitDepth );
+void processingSetBlackAndWhiteLevel(processingObject_t * processing,
+                                      float mlvBlackLevel, int mlvWhiteLevel, int mlvBitDepth );
 
 /* Get black/white level */
 #define processingGetBlackLevel(processing) (processing)->black_level
